requirements.txt
locust
openai
boto3
octoai-sdk